feat: refactor widget color handling and ID generation by consolidating functions
